break(button): remove the `asLink` prop
Determine if the button should be a link simply by
checking if there is a `href` given.

Also fix the bug where we had to give a `onClick` handler
for the  link to work.
And fix the style of a non text link button to remove
underlining

Finally fix the BEM name of the link dropdown item